# Use bisect for weighted choice implementation.