👓 @spladug @ckwang8128 

You nerd sniped me.  Couldn't stop thinking about it on the way home.