@dellis23 2017-01-04
If you guys like this I can update the docs to make the underlying search time more clear, as well as perhaps the generator exhaustion.