@pacejackson 2017-01-04
:no_bell: If we're worried about the performance of this, would it be worthwhile to use numpy for it?

https://docs.scipy.org/doc/numpy/reference/generated/numpy.random.choice.html