@ckwang8128 2017-01-04
I think this updated implementation is pretty solid and is basically what numpy does - generate CDF and then binary search (https://github.com/numpy/numpy/blob/b94c2b01ff7ef5b8dc44726512cfa232e9054882/numpy/random/mtrand/mtrand.pyx#L1017) - without a lot of overhead