@spladug 2017-01-04
I did some timing.

## Current master

(my naive linear implementation)

```
.....1000 iterations of picks from even weight lottery with 10 choices, 0.62μs per iteration
.....1000 iterations of picks from even weight lottery with 100 choices, 2.95μs per iteration
.....1000 iterations of picks from even weight lottery with 1000 choices, 28.01μs per iteration
.....1000 iterations of picks from even weight lottery with 100000 choices, 2777.58μs per iteration
```

## This branch

(dan's logarithmic implementation, massive improvement)

```
.....1000 iterations of picks from even weight lottery with 10 choices, 0.48μs per iteration
.....1000 iterations of picks from even weight lottery with 100 choices, 0.56μs per iteration
.....1000 iterations of picks from even weight lottery with 1000 choices, 0.62μs per iteration
.....1000 iterations of picks from even weight lottery with 100000 choices, 1.43μs per iteration
```

Doing `tuple` instead of `list` has no significant effect.

## Numpy

Oddly enough `numpy` does pretty poorly, even worse than the original linear one (please check my implementation here)

```
.....1000 iterations of picks from even weight lottery with 10 choices, 48.06μs per iteration
.....1000 iterations of picks from even weight lottery with 100 choices, 61.58μs per iteration
.....1000 iterations of picks from even weight lottery with 1000 choices, 167.05μs per iteration
.....1000 iterations of picks from even weight lottery with 100000 choices, 12475.12μs per iteration
```

I used this re-implementation of the `WeightedLottery` interface to test numpy: https://gist.github.com/spladug/80d0872b40a46bfd6d1ec1cc60214645 . It passes the distribution test, so it should provide correct results (but it fails some of the other tests because the `mock.patch` on `random.random` doesn't work on numpy)

## Methodology:

This test script was used unmodified for each test: https://gist.github.com/spladug/79ad34db3481a1373b87f017553aab5e