@spladug 2017-01-04
(note that linear was pretty quick in the ranges we'll be using it immediately, but it's good to see the logarithmic solution doing better at high volume)