@pacejackson 2017-01-04
If I had to guess, numpy probably rebuilds the underlying data structure each time you call it, so that would slow things down a lot as the list got bigger. Also, that build time probably isn't included in the timings for this branch. 

_Regardless_ though, this seems way better especially if your use case is same thing like "make this once when you start the app and use it some indeterminate number of times".