@spladug 2017-01-04
Yeah, that's definitely a possibility. If I move the `lottery = WeightedLottery(...)` construction line into the statement and out of setup, this branch gets way slower as expected:

```
.....1000 iterations of picks from even weight lottery with 10 choices, 3.89μs per iteration
.....1000 iterations of picks from even weight lottery with 100 choices, 20.72μs per iteration
.....1000 iterations of picks from even weight lottery with 1000 choices, 185.43μs per iteration
.....1000 iterations of picks from even weight lottery with 100000 choices, 17917.62μs per iteration
```