@dellis23 2017-01-04
Awesome discussion!  Thanks for doing that timing stuff @spladug :)